Revert "Phucdtd patch1.0"
